+++ internal/io/web/server.go
@@ -62,6 +62,7 @@ func Run(gnv gnverifier.GNverifier, port int) {
 	e.GET("/data_sources", dataSources(gnv))
 	e.GET("/data_sources/:id", dataSource(gnv))
 	e.GET("/name_strings/:id", nameString(gnv))
+	e.GET("/name_strings/widget/:id", nameStringWidget(gnv))
 	e.GET("/about", about(gnv))
 	e.GET("/api", api(gnv))
 
@@ -105,49 +106,11 @@ func api(gnv gnverifier.GNverifier) func(echo.Context) error {
 
 func nameString(gnv gnverifier.GNverifier) func(echo.Context) error {
 	return func(c echo.Context) error {
-		id, _ := url.QueryUnescape(c.Param("id"))
-		var ds []int
-		var allMatches bool
-		dsStr := c.QueryParam("data_sources")
-		if dsStr != "" {
-			dss := strings.Split(dsStr, ",")
-			for i := range dss {
-				num, err := strconv.Atoi(dss[i])
-				if err == nil {
-					ds = append(ds, num)
-				}
-			}
-		}
-		allMatches = c.QueryParam("all_matches") == "true"
-		inp := vlib.NameStringInput{
-			ID:             id,
-			DataSources:    ds,
-			WithAllMatches: allMatches,
-		}
-		res, err := gnv.NameString(inp)
+		data, err := getNameString(c, gnv)
 		if err != nil {
 			return err
 		}
-		var names []vlib.Name
-		if res.Name != nil {
-			names = []vlib.Name{*res.Name}
-		}
 
-		data := Data{
-			Input:         inp.ID,
-			Format:        "html",
-			DataSourceIDs: inp.DataSources,
-			AllMatches:    inp.WithAllMatches,
-			Page:          "home",
-			Verified:      names,
-			Version:       gnv.GetVersion().Version,
-		}
-		format := c.QueryParam("format")
-		if format == "csv" || format == "json" || format == "tsv" {
-			data.Format = format
-		}
-
-		fmt.Printf("NAME: %#v\n\n", data)
 		switch data.Format {
 		case "json":
 			return c.JSON(http.StatusOK, data.Verified)
@@ -161,6 +124,66 @@ func nameString(gnv gnverifier.GNverifier) func(echo.Context) error {
 			return c.Render(http.StatusOK, "layout", data)
 		}
 	}
+}
+
+func nameStringWidget(gnv gnverifier.GNverifier) func(echo.Context) error {
+	return func(c echo.Context) error {
+		data, err := getNameString(c, gnv)
+		if err != nil {
+			return err
+		}
+		return c.Render(http.StatusOK, "name_string_widget", data)
+	}
+}
+
+func getNameString(
+	c echo.Context,
+	gnv gnverifier.GNverifier,
+) (Data, error) {
+	var res Data
+	id, _ := url.QueryUnescape(c.Param("id"))
+	var ds []int
+	var allMatches bool
+	dsStr := c.QueryParam("data_sources")
+	if dsStr != "" {
+		dss := strings.Split(dsStr, ",")
+		for i := range dss {
+			num, err := strconv.Atoi(dss[i])
+			if err == nil {
+				ds = append(ds, num)
+			}
+		}
+	}
+	allMatches = c.QueryParam("all_matches") == "true"
+	inp := vlib.NameStringInput{
+		ID:             id,
+		DataSources:    ds,
+		WithAllMatches: allMatches,
+	}
+	out, err := gnv.NameString(inp)
+	if err != nil {
+		return res, err
+	}
+	var names []vlib.Name
+	if out.Name != nil {
+		names = []vlib.Name{*out.Name}
+	}
+
+	res = Data{
+		Input:         inp.ID,
+		Format:        "html",
+		DataSourceIDs: inp.DataSources,
+		AllMatches:    inp.WithAllMatches,
+		Page:          "home",
+		Verified:      names,
+		Version:       gnv.GetVersion().Version,
+	}
+	format := c.QueryParam("format")
+	if format == "csv" || format == "json" || format == "tsv" {
+		res.Format = format
+	}
+
+	return res, nil
 }
 
 func dataSources(gnv gnverifier.GNverifier) func(echo.Context) error {